add Node.js 18, 20 and 22 to `engines`
npm 10.8.2 doesn't install the `latest` tagged version of @pkgjs/support
with Node.js 22 as Node.js 22 is outside of the declared specification
for `engines`->`node`.

Regardless of whether the change in behaviour of npm is correct for a
semver-patch release of npm, update the `engines` field to include the
supported Node.js release lines at this point in time.